pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "coinclip"
version = "1.0.0"
description = "CoinCLIP: A Multimodal Framework for Assessing Memecoin Viability"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [{name = "tiramisu"}]

dependencies = [
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "transformers>=4.30.0",
    "pytorch-lightning>=2.0.0",
    "omegaconf>=2.3.0",
    "pillow>=9.5.0",
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "scikit-learn>=1.3.0",
    "tqdm>=4.65.0",
    "pyyaml>=6.0",
    "tensorboard>=2.13.0",
]

[project.scripts]
coinclip-prepare-data = "coinclip.scripts.prepare_data:main"
coinclip-train = "coinclip.scripts.train:main"
coinclip-evaluate = "coinclip.scripts.evaluate:main"

[tool.setuptools.packages.find]
where = ["src"]

